requirements.txt
pandas
twilio
python-dotenv
groq
requests